⚡️  Benchmarks scripts - genia & bigtext file
